refactor: move runtime plumbing into a dedicated hypervisor module
