+++ src/hypervisor/error.rs
