+++ src/hypervisor/mock.rs
@@ -2,7 +2,7 @@ use crate::api::dto::stats::InstanceStatsOutput;
 use crate::models::deployments::Deployment;
 use crate::models::health_check::{HealthCheck, HealthCheckStatus};
 use crate::models::volume::ResolvedMount;
-use crate::runtime::lifecycle_trait::{Log, RuntimeLifecycle};
+use crate::hypervisor::lifecycle_trait::{Log, RuntimeLifecycle};
 use async_trait::async_trait;
 use axum::response::sse::Event;
 use std::convert::Infallible;
