+++ src/hypervisor/types.rs
