+++ src/runtime/docker/docker_lifecycle.rs
@@ -2,7 +2,7 @@ use crate::api::dto::stats::InstanceStatsOutput;
 use crate::models::deployments::Deployment;
 use crate::models::health_check::HealthCheckStatus;
 use crate::models::volume::ResolvedMount;
-use crate::runtime::lifecycle_trait::{Log, RuntimeLifecycle, classify_log, extract_date};
+use crate::hypervisor::lifecycle_trait::{Log, RuntimeLifecycle, classify_log, extract_date};
 use crate::scheduler::intentional_shutdowns::IntentionalShutdowns;
 use async_trait::async_trait;
 use axum::response::sse::Event;
