+++ src/runtime/docker/lifecycle.rs
@@ -2,8 +2,8 @@ use super::container::{create_container, remove_container};
 use super::instances::list_instances;
 use crate::models::deployments::{Deployment, DeploymentStatus, MAX_RESTART_COUNT};
 use crate::models::volume::ResolvedMount;
-use crate::runtime::error::RuntimeError;
-use crate::runtime::types::InstanceStatus;
+use crate::hypervisor::error::RuntimeError;
+use crate::hypervisor::types::InstanceStatus;
 use crate::scheduler::intentional_shutdowns::IntentionalShutdowns;
 use bollard::Docker;
 use bollard::query_parameters::InspectContainerOptions;
